Hello,
This is Jamie White making a commit for Damballa.

I am committing a PRNG class that uses HMAC with an underlying hashing algorithm of SHA1. It is fairly straightforward, but seems to perform slightly better than some of the existing generators. Please let me know if I missed anything or did something incorrectly.

Thank you for the time to look at this.

- Jamie White